fix: list, view refactoring

+++ src/main/webapp/WEB-INF/views/snaps/list.jsp
@@ -6,6 +6,9 @@
 </head>
 <body>
 
+<%@ include file="/WEB-INF/views/header.jsp" %>
+
+
 <h1>Snap List</h1>
 
 <a href="write">✏️ 새 글 작성</a> |
@@ -13,9 +16,6 @@
 <a href="${pageContext.request.contextPath}/users/logout" style="float:right; color:red;">로그아웃</a>
 
 <hr/>
-
-<h1>Snap List</h1>
-<a href="write">새 글 작성</a> <hr/>
 
 <table border="1" width="80%">
     <tr>
@@ -30,11 +30,13 @@
         <tr>
             <td>${u.snap_id}</td>
             <td>${u.category}</td>
-            <td> <a href="view/${u.snap_id}">${u.snap_title}</a></td>
+            <td><a href="view/${u.snap_id}">${u.snap_title}</a></td>
             <td>${u.style}</td>
             <td>${u.price}</td>
             <td>
-                <a href="delete/${u.snap_id}">삭제</a>
+                <c:if test="${sessionScope.loginUser.user_id == u.user_id}">
+                    <a href="delete/${u.snap_id}" onclick="return confirm('정말 삭제하시겠습니까?');">삭제</a>
+                </c:if>
             </td>
         </tr>
     </c:forEach>
